[CHORE] #84 : 설문 확인 로직 management -> management/survey 모듈로 마이그레이션
